fix type

+++ src/unix.rs
@@ -286,7 +286,7 @@ impl MmapInner {
         let hugetlb = if huge.is_some() { MAP_HUGETLB } else { 0 };
         let hugetlb_size = huge.map_or(0, |mask| {
             (u64::from(mask) & (MAP_HUGE_MASK as u64)) << MAP_HUGE_SHIFT
-        });
+        }) as i32;
         MmapInner::new(
             len,
             libc::PROT_READ | libc::PROT_WRITE,
